% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battle_details.R
\docType{data}
\name{battle_details}
\alias{battle_details}
\title{Battle details}
\format{A data frame with 2 columns:
\describe{
  \item{battle}{character, battle name}
  \item{energy}{integer, battle energy cost}
}}
\usage{
battle_details
}
\description{
Battle details including energy cost
}
\keyword{datasets}
