% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battle_details.R
\docType{data}
\name{battle_rewards}
\alias{battle_rewards}
\title{Battle rewards}
\format{
A data frame with 2 columns:
\describe{
  \item{battle}{character, battle name}
  \item{reward}{character, name of reward}
}
}
\usage{
battle_rewards
}
\description{
Available rewards for each battle
}
\keyword{datasets}
