% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reward_details.R
\docType{data}
\name{reward_details}
\alias{reward_details}
\title{Reward details}
\format{
A data frame with 2 columns:
\describe{
  \item{reward}{character, reward name}
  \item{type}{character, reward type}
  \item{subtype}{character, reward subtype}
}
}
\usage{
reward_details
}
\description{
Reward details
}
\keyword{datasets}
