% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewards.R
\docType{data}
\name{rewards_raw}
\alias{rewards_raw}
\title{Raw rewards data}
\format{
A data.frame with 3 columns:
\describe{
  \item{battleID}{character, unique battle identifier}
  \item{reward}{character, name of reward}
  \item{count}{integer, amount of that reward received}
}
}
\usage{
rewards_raw
}
\description{
This data set is a concatenation of all the raw data files containing rewards 
received. This does not include any rewards that could have been received and 
therefore the \code{rewards} data set is most likely of more use.
}
\seealso{
\code{\link{rewards}}, \code{\link{battles}}, \code{\link{battle_rewards}}
}
\keyword{datasets}
