% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyApp.R
\name{runShinyApp}
\alias{runShinyApp}
\title{Run package shiny apps}
\usage{
runShinyApp(app)
}
\arguments{
\item{app}{The name of a shiny app in this package.}
}
\description{
Run package shiny apps
}
\details{
Launches a shiny app
}
