% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swgoh.R
\docType{package}
\name{swgoh}
\alias{swgoh}
\title{swgoh: An R package for Star Wars Galaxy of Heroes data}
\description{
This package contains data on battles fought and rewards obtained while 
playing the mobile game Star Wars Galaxy of Heroes.
}
\section{Data}{


\code{\link{rewards}}: this data set contains the rewards received from 
battles including details about the date the battle was fought, the user
who fought the battle, whether or not the battle was simulated, and, if 
simulated, the number of simulations run.
}

