% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\docType{data}
\name{users}
\alias{users}
\title{Users}
\format{A data frame with 3 columns:
\describe{
  \item{date}{Date, date of battle}
  \item{userID}{integer, user identifier}
  \item{level}{integer, player level (1-85)}
}}
\usage{
users
}
\description{
Level information for the three users for which data were taken
}
\keyword{datasets}
