% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewards.R
\docType{data}
\name{rewards}
\alias{rewards}
\title{Rewards for battles fought}
\format{
A data frame with 3 columns:
\describe{
  \item{date}{Date, date battle was fought}
  \item{battleID}{character, unique battle identifier using YYYYMMDD-### format}
  \item{userID}{character, unique account identifier}
  \item{battle}{character, battle name}
  \item{attempts}{integer, number of times battle was fought/simulated}
  \item{simulated}{logical, indicates if the battle was simulated (TRUE) or fought (FALSE)}
  \item{reward}{character, type of reward received}
  \item{count}{integer, number of that reward received}
}
}
\usage{
rewards
}
\description{
Data set that combines the data in \code{battles}, \code{battle_rewards},
and \code{rewards_raw}. This data set includes zeros for rewards not received
and therefore is more relevant than \code{rewards_raw}.
}
\seealso{
\code{\link{rewards_raw}}, \code{\link{battles}}, \code{\link{battle_rewards}}
}
\keyword{datasets}
