% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\docType{data}
\name{users}
\alias{users}
\title{Daily user data}
\format{A data frame with 4 columns:
\describe{
  \item{date}{Date, date of battle}
  \item{userID}{character, user identifier}
  \item{level}{integer, player's level (1-85)}
  \item{power}{integer, player's galactic power}
}}
\usage{
users
}
\description{
Daily level and galactic power for users
}
\keyword{datasets}
